requirements.txt
tensorflow-cpu
gunicorn
uwsgi
numpy
Pillow
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.3
streamlit==0.79.0
